requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
pydantic>=2.7.0,<3.0.0
alembic>=1.13.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0
